requirements.txt
pytest==7.2.1
selenium==4.8.0
webdriver-manage==3.8.5
pytest-html==3.2.0
pytest-xdist==3.2.0
